+++ server/app/controllers/userActions.js
@@ -69,7 +69,16 @@ const login = async (req, res, next) => {
 };
 
 // The D of BREAD - Destroy (Delete) operation
-// This operation is not yet implemented
+const destroy = async (req, res, next) => {
+  try {
+    const { id } = req.params;
+    await tables.user.delete(id);
+
+    res.sendStatus(204);
+  } catch (err) {
+    next(err);
+  }
+};
 
 // Ready to export the controller functions
 module.exports = {
@@ -77,6 +86,6 @@ module.exports = {
   read,
   // edit,
   add,
-  // destroy,
+  destroy,
   login,
 };
